+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - validation_tools module created
 - Defined is_valid_slot function for dynamic __slots__ creation.
 - Defined canonical_slots function for dynamic __slots__ creation.
+- Added the protocols module with Freeze and Freezable protocols.
 
 ### Changed
 - Update ruff and sync uv.lock